Data overview per corpus type
Based on the world map viewer, create a few more graphs that give more insight into the different corpus types. 

Ideally, this could also be used to quickly explore which kinds of gaps are being filled/still need filling through additional data-in. For that to be most useful, graphs that sub-divide by concept would also be good but these are not (yet) part of the open-data release.